refactor(kitchenpos): repository, 외부 API 요청 클래스 패키지 구조 이동
