+++ src/main/java/kitchenpos/products/infra/repository/JpaProductRepository.java
@@ -1,6 +1,7 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.infra.repository;
 
 import java.util.UUID;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
 import kitchenpos.products.tobe.domain.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
