+++ src/main/java/kitchenpos/products/tobe/domain/model/DisplayedName.java
@@ -6,7 +6,7 @@ import static kitchenpos.common.exception.ExceptionDetails.DISPLAYED_NAME_INCLUD
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
-import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.service.PurgomalumClient;
 import kitchenpos.products.tobe.domain.exception.DisplayedNameEmptyException;
 import kitchenpos.products.tobe.domain.exception.DisplayedNameIncludeProfanityException;
