+++ src/main/java/kitchenpos/products/tobe/domain/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.tobe.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
