+++ src/main/java/kitchenpos/products/tobe/domain/service/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.infra;
+package kitchenpos.products.tobe.domain.service;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
