+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.application;
 
 import kitchenpos.products.tobe.domain.model.Product;
-import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
