+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -18,8 +18,8 @@ import kitchenpos.products.application.dto.CreateProductRequestDto;
 import kitchenpos.products.application.dto.CreateProductResponseDto;
 import kitchenpos.products.application.dto.FindProductResponseDto;
 import kitchenpos.products.application.service.ProductService;
-import kitchenpos.products.domain.ProductRepository;
-import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
+import kitchenpos.products.tobe.domain.service.PurgomalumClient;
 import kitchenpos.products.tobe.domain.model.Product;
 import kitchenpos.products.tobe.domain.model.ProductPrice;
 import org.junit.jupiter.api.BeforeEach;
