+++ src/test/java/kitchenpos/products/domain/DisplayedNameTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import kitchenpos.products.application.FakePurgomalumClient;
-import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.products.tobe.domain.service.PurgomalumClient;
 import kitchenpos.products.tobe.domain.exception.DisplayedNameEmptyException;
 import kitchenpos.products.tobe.domain.exception.DisplayedNameIncludeProfanityException;
 import kitchenpos.products.tobe.domain.model.DisplayedName;
